Added a function to add callbacks without menu items and made the Sender of the channel publicly available
That way other threads can trigger callbacks of the systray Application independently from menu
items and user iteraction. I needed this feature to tell the Application from another thread
that it should update it's icon. The Sender "tx" is inside of an Option<_> so the Application can
"drop" it by overwriting it with None when  quit or shutdown is triggered.